+++ README.rst
@@ -1,7 +1,11 @@
-packaging
-=========
+packaging2
+==========
 
 .. start-intro
+
+packaging2 is friendly fork of packaging to work around the drop for
+LegacyVersion support. See https://github.com/pypa/packaging/issues/530
+
 
 Reusable core utilities for various Python Packaging
 `interoperability specifications <https://packaging.python.org/specifications/>`_.
@@ -12,7 +16,7 @@ or benefit greatly from having a single shared implementation (eg: :pep:`425`).
 
 .. end-intro
 
-The ``packaging`` project includes the following: version handling, specifiers,
+The ``packaging2`` project includes the following: version handling, specifiers,
 markers, requirements, tags, utilities.
 
 Documentation
@@ -32,18 +36,15 @@ Installation
 
 Use ``pip`` to install these utilities::
 
-    pip install packaging
+    pip install packaging2
 
 Discussion
 ----------
 
 If you run into bugs, you can file them in our `issue tracker`_.
 
-You can also join ``#pypa`` on Freenode to ask questions or get involved.
 
-
-.. _`documentation`: https://packaging.pypa.io/
-.. _`issue tracker`: https://github.com/pypa/packaging/issues
+.. _`issue tracker`: https://github.com/nexB/packaging2/issues
 
 
 Code of Conduct
@@ -61,13 +62,9 @@ The ``CONTRIBUTING.rst`` file outlines how to contribute to this project as
 well as how to report a potential security issue. The documentation for this
 project also covers information about `project development`_ and `security`_.
 
-.. _`project development`: https://packaging.pypa.io/en/latest/development/
-.. _`security`: https://packaging.pypa.io/en/latest/security/
 
 Project History
 ---------------
 
 Please review the ``CHANGELOG.rst`` file or the `Changelog documentation`_ for
 recent changes and project history.
-
-.. _`Changelog documentation`: https://packaging.pypa.io/en/latest/changelog/
