+++ docs/utils.rst
@@ -1,7 +1,7 @@
 Utilities
 =========
 
-.. currentmodule:: packaging.utils
+.. currentmodule:: packaging2.utils
 
 
 A set of small, helper utilities for dealing with Python packages.
@@ -19,7 +19,7 @@ Reference
 
     .. doctest::
 
-        >>> from packaging.utils import canonicalize_name
+        >>> from packaging2.utils import canonicalize_name
         >>> canonicalize_name("Django")
         'django'
         >>> canonicalize_name("oslo.concurrency")
@@ -30,14 +30,14 @@ Reference
 .. function:: canonicalize_version(version)
 
     This function takes a string representing a package version (or a
-    :class:`~packaging.version.Version` instance), and returns the
+    :class:`~packaging2.version.Version` instance), and returns the
     normalized form of it.
 
     :param str version: The version to normalize.
 
     .. doctest::
 
-        >>> from packaging.utils import canonicalize_version
+        >>> from packaging2.utils import canonicalize_version
         >>> canonicalize_version('1.4.0.0.0')
         '1.4'
 
@@ -47,7 +47,7 @@ Reference
     returning a tuple of name, version, build number, and tags.
 
     The name part of the tuple is normalized. The version portion is an
-    instance of :class:`~packaging.version.Version`. The build number
+    instance of :class:`~packaging2.version.Version`. The build number
     is ``()`` if there is no build number in the wheel filename,
     otherwise a two-item tuple of an integer for the leading digits and
     a string for the rest of the build number. The tags portion is an
@@ -57,9 +57,9 @@ Reference
 
     .. doctest::
 
-        >>> from packaging.utils import parse_wheel_filename
-        >>> from packaging.tags import Tag
-        >>> from packaging.version import Version
+        >>> from packaging2.utils import parse_wheel_filename
+        >>> from packaging2.tags import Tag
+        >>> from packaging2.version import Version
         >>> name, ver, build, tags = parse_wheel_filename("foo-1.0-py3-none-any.whl")
         >>> name
         'foo'
@@ -75,14 +75,14 @@ Reference
     This function takes the filename of a sdist file (as specified
     in the `Source distribution format`_ documentation), and parses
     it, returning a tuple of the normalized name and version as
-    represented by an instance of :class:`~packaging.version.Version`.
+    represented by an instance of :class:`~packaging2.version.Version`.
 
     :param str filename: The name of the sdist file.
 
     .. doctest::
 
-        >>> from packaging.utils import parse_sdist_filename
-        >>> from packaging.version import Version
+        >>> from packaging2.utils import parse_sdist_filename
+        >>> from packaging2.version import Version
         >>> name, ver = parse_sdist_filename("foo-1.0.tar.gz")
         >>> name
         'foo'
