+++ packaging2/_structures.py
