+++ setup.py
@@ -21,21 +21,15 @@ except ImportError:
 base_dir = os.path.dirname(__file__)
 
 about = {}
-with open(os.path.join(base_dir, "packaging", "__about__.py")) as f:
+with open(os.path.join(base_dir, "packaging2", "__about__.py")) as f:
     exec(f.read(), about)
 
 with open(os.path.join(base_dir, "README.rst")) as f:
     long_description = f.read()
 
 with open(os.path.join(base_dir, "CHANGELOG.rst")) as f:
-    # Remove :issue:`ddd` tags that breaks the description rendering
-    changelog = re.sub(
-        r":issue:`(\d+)`",
-        r"`#\1 <https://github.com/pypa/packaging/issues/\1>`__",
-        f.read(),
-    )
+    changelog = f.read()
     long_description = "\n".join([long_description, changelog])
-
 
 setup(
     name=about["__title__"],
@@ -65,6 +59,6 @@ setup(
         "Programming Language :: Python :: Implementation :: CPython",
         "Programming Language :: Python :: Implementation :: PyPy",
     ],
-    packages=["packaging"],
-    package_data={"packaging": ["py.typed"]},
+    packages=["packaging2"],
+    package_data={"packaging2": ["py.typed"]},
 )
