+++ tests/test_tags.py
@@ -19,8 +19,8 @@ import types
 import pretend
 import pytest
 
-from packaging import tags
-from packaging._musllinux import _MuslVersion
+from packaging2 import tags
+from packaging2._musllinux import _MuslVersion
 
 
 @pytest.fixture
